Finally it builds...but nothing on screen yet :(

+++ WindowsGame1/WindowsGame1/Components/Velocity.cs
@@ -51,9 +51,9 @@ namespace StarWarrior.Components
             angle = (angle + a) % 360;
         }
 
-        public double GetAngleAsRadians()
+        public float GetAngleAsRadians()
         {
-            return Math.PI * angle / 180.0; ;
+            return (float)Math.PI * angle / 180.0f; ;
         }
     }
 }
